backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
psycopg2-binary==2.9.10
alembic==1.16.5
sqlalchemy==2.0.43
python-multipart==0.0.20
pydantic==2.11.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
boto3==1.35.0
celery==5.3.6
redis==5.0.1
requests==2.31.0
cryptography==41.0.7
pillow==10.0.1
